fix(anidb): official XML text goes through the hardened markup pipeline
The XML parser grew its own tag strip and entity decoder, both weaker than the
ones the browse scraper already had, and CodeQL flagged the tag strip.

Two real holes, not just a lint:

- `value.replace(/<[^>]+>/g, "")` is a single pass, so
  `<<script>script>alert(1)</script>` leaves a live `<script` behind, and on a
  body of many `<` with no `>` it is quadratic — this runs on fetched provider
  markup.
- The XML entity decoder happily minted control characters. `&#27;` is plain
  ASCII in the response body, so an official summary could hand the terminal a
  live ESC byte — cursor moves, screen clears, OSC 52 clipboard writes — with
  no raw control byte ever crossing the wire. Episode titles and synopses go
  straight to stdout.

The browse parser had already closed both, so its scanners now live in
`shared/markup-text` as one `markupToPlainText` pipeline that both the HTML
scraper and the XML parser use: script spans, then tags, entities decoded
exactly once, control characters dropped, whitespace collapsed.

The two remaining CodeQL alerts were test stubs matching
`url.includes("api.anidb.net:9001")`, which would answer any URL merely
containing that text — including a request the provider should never have made.
They now match host and port.
